requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
frozenlist==1.5.0
future==1.0.0
greenlet==3.2.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-core==0.3.52
langchain-openai==0.3.13
langchain-text-splitters==0.3.8
langsmith==0.3.31
line-bot-sdk==3.16.3
MarkupSafe==3.0.2
multidict==6.4.3
openai==1.74.0
orjson==3.10.16
packaging==24.2
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
